refactor(catalog-tests): move test base files to Base folder and fix fixture lifecycle
