+++ src/Ecommerce.Catalog/Ecommerce.Catalog.IntegrationTests/Base/CatalogDatabaseConfiguration.cs
@@ -1,6 +1,6 @@
 using Ecommerce.Shared.IntegrationTests.Database;
 
-namespace Ecommerce.Catalog.IntegrationTests;
+namespace Ecommerce.Catalog.IntegrationTests.Base;
 
 public sealed class CatalogDatabaseConfiguration(string connectionString) : IDatabaseConfiguration
 {
