+++ src/Ecommerce.Catalog/Ecommerce.Catalog.IntegrationTests/Base/CatalogTestCollection.cs
@@ -1,0 +1,4 @@
+namespace Ecommerce.Catalog.IntegrationTests.Base;
+
+[CollectionDefinition(nameof(CatalogTestCollection))]
+public class CatalogTestCollection : ICollectionFixture<CatalogIntegrationFixture>;
