+++ src/Ecommerce.Catalog/Ecommerce.Catalog.IntegrationTests/Base/CatalogWebApplicationFactory.cs
@@ -1,7 +1,7 @@
 using Ecommerce.Shared.IntegrationTests;
 using Ecommerce.Shared.IntegrationTests.Database;
 
-namespace Ecommerce.Catalog.IntegrationTests;
+namespace Ecommerce.Catalog.IntegrationTests.Base;
 
 public sealed class CatalogWebApplicationFactory(DatabaseContainerFixture containerFixture)
     : EcommerceWebApplicationFactory(new CatalogDatabaseConfiguration(containerFixture.ConnectionString));
