--- src/Ecommerce.Catalog/Ecommerce.Catalog.IntegrationTests/CatalogTestCollection.cs
@@ -1,8 +1,0 @@
-using Ecommerce.Shared.IntegrationTests.Database;
-
-namespace Ecommerce.Catalog.IntegrationTests;
-
-[CollectionDefinition(nameof(CatalogTestCollection))]
-public class CatalogTestCollection
-    : ICollectionFixture<DatabaseContainerFixture>,
-      ICollectionFixture<CatalogIntegrationFixture>;
